Restructure runtime and worker, create files for materializer service
