+++ aquadoggo/src/lib.rs
@@ -13,20 +13,20 @@
     unused_qualifications
 )]
 
+mod bus;
 mod config;
 mod context;
 mod db;
 mod errors;
 mod graphql;
+mod manager;
+mod materializer;
+mod node;
 mod rpc;
-mod runtime;
 mod server;
-mod service_manager;
-mod service_message;
-mod worker;
 
 #[cfg(test)]
 mod test_helpers;
 
 pub use config::Configuration;
-pub use runtime::Runtime;
+pub use node::Node;
