+++ aquadoggo/src/manager.rs
