+++ aquadoggo/src/materializer/mod.rs
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+mod worker;
+mod tasks;
