+++ aquadoggo/src/materializer/tasks/mod.rs
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+mod reduce;
+mod dependencies;
+mod schema;
