+++ aquadoggo/src/materializer/tasks/reduce.rs
@@ -1,0 +1,1 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
