+++ aquadoggo/src/materializer/worker.rs
